+++ web/apps/accounts/src/types/context.ts
@@ -12,7 +12,7 @@ type AppContextT = Omit<AccountsContextT, "logout">;
 export const AppContext = createContext<AppContextT | undefined>(undefined);
 
 /**
- * Utility hook to get the {@link AppContextT}, throwing an exception if it is
- * not defined.
+ * Utility hook to get the {@link AppContextT} expected to be available to all
+ * React components in the Accounts app's React tree.
  */
 export const useAppContext = (): AppContextT => useContext(AppContext)!;
