+++ web/apps/auth/src/pages/change-email.tsx
@@ -1,5 +1,5 @@
 import Page_ from "@/accounts/pages/change-email";
-import { useAppContext } from "./_app";
+import { useAppContext } from "types/context";
 
 const Page = () => <Page_ appContext={useAppContext()} />;
