+++ web/apps/auth/src/pages/login.tsx
@@ -1,5 +1,5 @@
 import Page_ from "@/accounts/pages/login";
-import { useAppContext } from "./_app";
+import { useAppContext } from "types/context";
 
 const Page = () => <Page_ appContext={useAppContext()} />;
