+++ web/apps/auth/src/pages/passkeys/finish.tsx
@@ -1,5 +1,5 @@
 import Page_ from "@/accounts/pages/passkeys/finish";
-import { useAppContext } from "../_app";
+import { useAppContext } from "types/context";
 
 const Page = () => <Page_ appContext={useAppContext()} />;
