+++ web/apps/auth/src/pages/two-factor/recover.tsx
@@ -1,5 +1,5 @@
 import Page_ from "@/accounts/pages/two-factor/recover";
-import { useAppContext } from "../_app";
+import { useAppContext } from "types/context";
 
 const Page = () => <Page_ appContext={useAppContext()} twoFactorType="totp" />;
